Add listener for "windows resize"
Now we can subscribe for the event of windows resize and change UI settings accordingly to the size of the windows for example. I added to event.

The first one is just a wrapper for the domready event of jquery (to be consistent). We can subscribe to this event like this:

		// When the dom is ready, update the UI:
		radio("domready").subscribe(resize);

The second one is fired when the windows has been resize:

		// When the windows is resized, change UI
		radio("window:resize").subscribe(resize);